+++ src/picounits/__init__.py
@@ -1,24 +1,27 @@
 # pylint: skip-file
 # picounits/__init__.py
 
-from typing import Any
-
-from picounits.extensions.parser import Parser
+from picounits.extensions.parser import Parser, resolve_derived
 from picounits.extensions.loader import DynamicLoader
 
 from picounits.constants import *
-from picounits.core.quantities.validator import expects
+from picounits.utilities.validation import expects, strip_quantity, check_quantity
 from picounits.core.quantities.packet import Packet as Quantity
-from picounits.configuration.management import reload_config
 
-# Reloads the users .picounits configuration file.
-reload_config()
+from picounits.configuration.management import inject_unit_frame
+from picounits.utilities.errors import UnitError
+
+
+# Configuration / Management for forcing a unit frame at the application level.
+_ = inject_unit_frame
 
 # References for quantities when doing type hinting.
 Q = Quantity
 q = Quantity
 
-_ = expects 
+# Validation
+_, _ = strip_quantity, check_quantity
+_ = UnitError
 
 # LEGACY API - keep the old name for backward compatibility before 1.0.6
 unit_validator = expects
@@ -26,45 +29,7 @@ unit_validator = expects
 # Parser & Loader import
 _ = Parser
 _ = DynamicLoader
-
-
-class UnitError(TypeError):
-    """ Exception for Unit Error """
-    def __init__(self, error: str, messenger: str |  None = None):
-        """ Returns a custom error message """
-        if messenger:
-            msg = f"{messenger!r} raised error: {error}."
-        else:
-            msg = f"Unit error occurred: {error}."
-        super().__init__(msg)
-
-
-def check_quantity(quantity: Quantity, ref: Quantity) -> None:
-    """ Checks if the quantity has the correct reference unit """
-    if not isinstance(quantity, Quantity):
-        msg = f"{type(quantity)!r} is not a physical quantity object"
-        raise UnitError(msg)
-
-    if not isinstance(ref, (Unit, Quantity)):
-        msg = f"Reference unit must be either a quantity or unit, not {type(ref)}"
-        raise UnitError(msg)
-
-    if isinstance(ref, Quantity):
-        if quantity.unit != ref.unit:
-            msg = f"Expected {ref.unit!r}, got {quantity.unit!r}"
-            raise UnitError(msg)
-
-    if isinstance(ref, Unit):
-        if quantity.unit != ref:
-            msg = f"Expected {ref!r}, got {quantity.unit!r}"
-            raise UnitError(msg)
-
-
-def strip_quantity(quantity: Quantity, reference: Quantity) -> Any:
-    """ Strips quantity from value returns raw value """
-    check_quantity(quantity, reference)
-
-    return quantity.value
+_ = resolve_derived
 
 
 # API Promises
@@ -78,7 +43,9 @@ __all__ = [
     "Quantity",
     "Q",
     "q",
-    "expects"
+    "expects",
+    "inject_unit_frame",
+    "ResolveDerived"
     
     # Scales
     "GIGA", "giga",