+++ src/picounits/core/quantities/packet.py
@@ -16,7 +16,7 @@ from typing import Any
 from picounits.core.scales import PrefixScale
 from picounits.core.unit import Unit
 
-from picounits.lazy_imports import import_factory
+from picounits.utilities.lazy_imports import import_factory
 
 
 @dataclass
@@ -24,7 +24,8 @@ class Packet(ABC):
     """
     An Abstract Physical Packet: A Prefix, Value and Unit
 
-    NOTE: Prefix is init-only (InitVar)
+    NOTE: 
+    Prefix is init-only (InitVar)
     All packets, normalize value to BASE unit during init
     """
     value: Any