+++ src/picounits/core/quantities/scalars/scalar.py
@@ -23,18 +23,20 @@ class ScalarPacket(Packet, ABC):
     """
     An Abstract Scalar Packet: A prefix, value (Any Scalar) and Unit
 
-    NOTE: Representation, prefix scaling, comparison, validation
+    NOTE:
+    Representation, prefix scaling, comparison, validation
     are not implemented in this base case.
     """
-
     def sqrt(self) -> Packet:
         """ Defines the behavior for taking the square root of a scalar """
-        q2 = self._get_other_packet(1 / 2)  # Due to fractional exponent law
+        # Due to fractional exponent law
+        q2 = self._get_other_packet(1 / 2)
         return acops.power_logic(self, q2)
 
     def cbrt(self) -> Packet:
         """ Defines the behavior for taking the cubic root of a scalar """
-        q2 = self._get_other_packet(1 / 3)  # Due to fractional exponent law
+        # Due to fractional exponent law
+        q2 = self._get_other_packet(1 / 3)
         return acops.power_logic(self, q2)
 
     def __add__(self, other: Any) -> Packet: