+++ src/picounits/core/quantities/scalars/types/complex.py
@@ -20,16 +20,14 @@ from picounits.core.scales import PrefixScale
 from picounits.core.quantities.packet import Packet
 from picounits.core.quantities.scalars.scalar import ScalarPacket
 
-from picounits.lazy_imports import import_factory
-from picounits.configuration.picounits import DEFAULT_SIGNIFICANT_FIGURES
+from picounits.utilities.lazy_imports import import_factory
+from picounits.configuration.management import get_significant_figures
 
 
 @dataclass(slots=True, repr=False, unsafe_hash=True)
 class ComplexPacket(ScalarPacket):
     """
     A Complex Packet: A prefix, value (Real + Imaginary) and Unit
-
-    NOTE: Prefix is init-only, value is held in absolute form
     """
     def __post_init__(self, prefix: PrefixScale) -> None:
         """ Validates value and unit, then mutates value to Base """
@@ -61,11 +59,11 @@ class ComplexPacket(ScalarPacket):
 
         if isinstance(value, complex):
             # Rounds imaginary & real parts if complex
-            real = round(value.real, DEFAULT_SIGNIFICANT_FIGURES)
-            imag = round(value.imag, DEFAULT_SIGNIFICANT_FIGURES)
+            real = round(value.real, get_significant_figures())
+            imag = round(value.imag, get_significant_figures())
             value = complex(real, imag)
         else:
-            value = round(value, DEFAULT_SIGNIFICANT_FIGURES)
+            value = round(value, get_significant_figures())
 
         return f"{value} {prefix}({self.unit.name(fundamental)})"
 
@@ -154,7 +152,7 @@ class ComplexPacket(ScalarPacket):
 
         if not format_spec:
             # Fall-back if no formatting is provided in the f-string
-            format_spec = f".{DEFAULT_SIGNIFICANT_FIGURES}f"
+            format_spec = f".{get_significant_figures()}f"
 
         formatted_value = format(value, format_spec)
         return f"{formatted_value} {prefix}({self.unit.name()})"