+++ src/picounits/utilities/lazy_imports.py
@@ -7,20 +7,10 @@ Description:
     factory as its the main lazy import.
 """
 
-
 from typing import Any
 from functools import lru_cache
 
-
-class LazyImportError(ImportError):
-    """ Exception for failed lazy imports """
-    def __init__(self, caller: str, module: str):
-        """ Returns a custom error message for lazy imports """
-        msg = (
-            f"Could not import '{module}' for '{caller}'. "
-            "This usually means picounits was not installed correctly"
-        )
-        super().__init__(msg)
+from picounits.utilities.errors import LazyImportError
 
 
 @lru_cache(maxsize=None)
@@ -35,9 +25,7 @@ def import_factory(caller_name: str) -> Any:
 
 
 @lru_cache(maxsize=None)
-def lazy_import(
-    module_path: str, method_name: str, caller_name: str
-) -> Any:
+def lazy_import(module_path: str, method_name: str, caller_name: str) -> Any:
     """ Caches/returns the module_path.module_name for lazy imports  """
     try:
         mod = __import__(module_path, fromlist=[method_name])